refactor: auth queryKey, option query 폴더로 이사
