+++ apps/web/src/entities/auth/config/index.ts
@@ -1,3 +1,2 @@
 export * from './constants';
 export * from './query';
-export * from './query-key';
