--- apps/web/src/entities/auth/config/query-key/index.ts
@@ -1,1 +1,0 @@
-export * from './authQueryKey';
