+++ apps/web/src/entities/auth/config/query/authQueryKey.ts
