+++ apps/web/src/entities/auth/config/query/authQueryOptions.ts
@@ -3,7 +3,8 @@ import { queryOptions } from '@tanstack/react-query';
 import { QUERY_STALE_TIME } from '@/shared/constants';
 
 import { getAdmissionState, getMyInfo } from '../../api';
-import { authQueryKey } from '../query-key';
+
+import { authQueryKey } from './authQueryKey';
 
 export const authQueryOptions = {
   me: () =>
