+++ apps/web/src/entities/auth/config/query/index.ts
@@ -1,1 +1,2 @@
+export { authQueryKey } from './authQueryKey';
 export { authQueryOptions } from './authQueryOptions';
